Add stand-in communication channels between webview and extension context.
The webview can call Review/Explain commands on button click.
The extension context forces a webview refresh on successful model response.

There are definitely some things that are temporary.
Some types/interfaces are duplicated in webview and extension src/, which is not great.

Response content from the model is not formatted in the webview.